Move metadata to restate-core
as described in the previous PR in the stack
