+++ crates/core/Cargo.toml
@@ -14,7 +14,9 @@ test-util = []
 restate-types = { workspace = true }
 
 anyhow = { workspace = true }
+arc-swap = { workspace = true }
 derive_more = { workspace = true }
+enum-map = { workspace = true }
 futures = { workspace = true }
 static_assertions = { workspace = true }
 strum = { workspace = true }
