+++ crates/core/src/lib.rs
@@ -8,8 +8,10 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod task_center_types;
-pub use task_center_types::*;
+pub mod metadata;
 
 mod task_center;
+mod task_center_types;
+
 pub use task_center::*;
+pub use task_center_types::*;
