+++ crates/types/src/lib.rs
@@ -13,8 +13,8 @@
 mod base62_util;
 mod id_util;
 mod macros;
-mod metadata;
 mod node_id;
+mod version;
 
 pub mod deployment;
 pub mod errors;
@@ -31,5 +31,5 @@ pub mod time;
 pub mod timer;
 
 pub use id_util::{IdDecoder, IdEncoder, IdResourceType, IdStrCursor};
-pub use metadata::*;
 pub use node_id::*;
+pub use version::*;
